+++ OliveDI/src/test/java/olive/di/fixture/InstanceCreateTestFixtures.kt
@@ -1,0 +1,4 @@
+package olive.di.fixture
+
+class Foo
+
