+++ OliveDI/src/test/java/olive/di/fixture/RecursiveDITestFixtures.kt
@@ -1,0 +1,5 @@
+package olive.di.fixture
+
+class Sandwich(olive: Olive)
+
+class Olive
